+++ src/analysis/models/order_type.py
@@ -1,0 +1,6 @@
+from enum import Enum
+
+
+class OrderType(Enum):
+    Increment = 0
+    Decrement = 1
