+++ src/analysis/models/similarity_type.py
@@ -31,8 +31,3 @@ class SimilarityType(Enum):
     def get_all_items():
         return [SimilarityType.TextEdit, SimilarityType.Direct, SimilarityType.Consistency,
                 SimilarityType.Pattern, SimilarityType.Blat]
-
-
-class OrderType(Enum):
-    Increment = 0
-    Decrement = 1
