+++ src/analysis/similarities/consistency_similarity.py
@@ -1,7 +1,34 @@
+from typing import Tuple
+
 from analysis.similarities.base_similarity import BaseSimilarity
 
 
 class ConsistencySimilarity(BaseSimilarity):
+    def rendering_sequence(self, gene: str, database: str, offset: int) -> Tuple[list, list, list]:
+        sequence_gene = []
+        sequence_target = []
+        sequence = []
+        tot = len(gene)
+        score, (score_queue, score_merge_idx) = self.get_similarity(gene, database, offset)
+        sequence_gene.extend(gene[:])
+        sequence_target.extend(database[offset:offset + tot])
+        cur_pos = 0
+        for idx, (same_cnt, same_end) in enumerate(score_queue):
+            same_start = same_end - same_cnt
+            while cur_pos < same_start:
+                if score_merge_idx[0] < idx <= score_merge_idx[1]:
+                    sequence.append('-')
+                else:
+                    sequence.append('.')
+                cur_pos += 1
+            while cur_pos < same_end:
+                sequence.append('*')
+                cur_pos += 1
+        while cur_pos < tot:
+            sequence.append('.')
+            cur_pos += 1
+        return sequence_gene, sequence_target, sequence
+
     def __init__(self, max_patience: int):
         self.max_patience = max_patience
